pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gai"
version = "0.1.0"
authors = [
    { name = "Muzahid", email = "mmsaddam@gmail.com" },
]
description = "An AI-powered git commit message generator."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Utilities",
]
dependencies = [
    "requests",
    "python-dotenv",
    "openai",
]

[project.urls]
"Homepage" = "https://github.com/muzahid59/gai"
"Bug Tracker" = "https://github.com/muzahid59/gai/issues"

[project.scripts]
gai = "gai.cli:main"
